% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagrangeParametric.R
\name{lagrange_parametric}
\alias{lagrange_parametric}
\title{lagrange_parametric}
\usage{
lagrange_parametric(t, x, y)
}
\arguments{
\item{x}{Vector y Vector t Vector. Los tres vectores deben tener el mismo numero de elementos.}
}
\description{
Esta funcion devuelve las ecuaciones \code{x(t)} y \code{y(t)} para una curva parametrica.
}
\details{
Esta funcion es parte del paquete de analisis numerico creado por los estudiantes de la
Pontificia Universidad Javeriana para este curso en el semestre 1830.
}
\examples{
t = c(0,0.25,0.5,0.75,1)
x = c(-1,0,1,0,1)
y = c(0,1,0.5,0,-1)
formulas = lagrange_parametric(t,x,y)
}
